Use Xorshift64starHasher for SMBIOS table modification detection
Move Xorshift64starHasher from patina_dxe_core to a shared
patina::hash module and replace the simple wrapping-add checksum in
SmbiosManager with the hasher. This avoids false negatives when
multiple byte changes cancel each other out (e.g. FieldA: 1→0 and
FieldB: 0→1).

Closes #1161
