+++ patina_dxe_core/src/protocol_db.rs
@@ -16,7 +16,7 @@ use alloc::{
     vec::Vec,
 };
 use core::{cmp::Ordering, ffi::c_void, hash::Hasher};
-use patina::error::EfiError;
+use patina::{error::EfiError, hash::Xorshift64starHasher};
 use r_efi::efi;
 
 use crate::tpl_mutex;
@@ -876,48 +876,6 @@ impl SpinLockedProtocolDb {
 
 unsafe impl Send for SpinLockedProtocolDb {}
 unsafe impl Sync for SpinLockedProtocolDb {}
-
-/// A hasher that uses the Xorshift64* algorithm to generate a random number to xor with the input bytes.
-///
-/// https://en.wikipedia.org/wiki/Xorshift#xorshift*
-struct Xorshift64starHasher {
-    state: u64,
-}
-
-impl Xorshift64starHasher {
-    /// Initialize the hasher with a seed.
-    fn new(seed: u64) -> Self {
-        Xorshift64starHasher { state: seed }
-    }
-
-    /// Generate a new random state.
-    fn next_state(&mut self) -> u64 {
-        self.state ^= self.state >> 12;
-        self.state ^= self.state << 25;
-        self.state ^= self.state >> 27;
-        self.state = self.state.wrapping_mul(0x2545F4914F6CDD1D);
-        self.state
-    }
-}
-
-impl Default for Xorshift64starHasher {
-    fn default() -> Self {
-        Xorshift64starHasher::new(compile_time::unix!())
-    }
-}
-
-impl Hasher for Xorshift64starHasher {
-    fn finish(&self) -> u64 {
-        self.state
-    }
-
-    fn write(&mut self, bytes: &[u8]) {
-        for &byte in bytes {
-            self.state ^= byte as u64;
-            self.state = self.next_state();
-        }
-    }
-}
 
 #[cfg(test)]
 #[coverage(off)]
@@ -2164,30 +2122,5 @@ mod tests {
                 assert!(child_list.contains(&child));
             }
         });
-    }
-
-    #[test]
-    fn xorshift64starhasher_test_different_seeds() {
-        let seed1 = 12345;
-        let seed2 = 54321;
-        let mut hasher1 = Xorshift64starHasher::new(seed1);
-        let mut hasher2 = Xorshift64starHasher::new(seed2);
-
-        let num1 = hasher1.next_state();
-        let num2 = hasher2.next_state();
-
-        assert_ne!(num1, num2, "Random numbers should be different for different seeds");
-    }
-
-    #[test]
-    fn xorshift64starhasher_test_same_seed() {
-        let seed = 12345;
-        let mut hasher1 = Xorshift64starHasher::new(seed);
-        let mut hasher2 = Xorshift64starHasher::new(seed);
-
-        let num1 = hasher1.next_state();
-        let num2 = hasher2.next_state();
-
-        assert_eq!(num1, num2, "Random numbers should be the same for the same seed");
     }
 }
