+++ sdk/patina/Cargo.toml
@@ -30,6 +30,7 @@ patina_macro = { workspace = true }
 fixedbitset = { workspace = true }
 
 cfg-if = { workspace = true }
+compile-time = { workspace = true }
 goblin = { workspace = true, features = ["pe32", "pe64", "te", "alloc"]}
 log = { workspace = true }
 r-efi = { workspace = true }
